Made changes Robert suggested and added Robert's solution at the bottom of HotButton.tsx for comparsion
